// src/me/onenote/resources/item/onenoteResourceItemRequestBuilderDeleteRequestConfiguration.ts
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface OnenoteResourceItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}